Rename from RePositionSpawnable to PositionSpawnable
